+++ src/components/icons/AgentIcons.tsx
@@ -1,0 +1,38 @@
+import React from 'react';
+import { SvgIcon, SvgIconProps } from '@mui/material';
+
+export const SeldonIcon = (props: SvgIconProps) => (
+  <SvgIcon {...props} viewBox="0 0 24 24">
+    <path d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm0 2c4.41 0 8 3.59 8 8s-3.59 8-8 8-8-3.59-8-8 3.59-8 8-8zm1 3v4.17l3.41 3.42-1.41 1.41L11 12V7h2z" />
+  </SvgIcon>
+);
+
+export const BaleyIcon = (props: SvgIconProps) => (
+  <SvgIcon {...props} viewBox="0 0 24 24">
+    <path d="M13 3c3.88 0 7 3.12 7 7 0 2.05-.89 3.9-2.29 5.18l3.29 3.29-1.41 1.41-3.29-3.29A6.995 6.995 0 0113 17c-3.88 0-7-3.12-7-7s3.12-7 7-7m0 2c-2.76 0-5 2.24-5 5s2.24 5 5 5 5-2.24 5-5-2.24-5-5-5zM12 8v3.5l2.5 1.5-.75 1.3L11 12.5V8h1z" />
+  </SvgIcon>
+);
+
+export const DorsIcon = (props: SvgIconProps) => (
+  <SvgIcon {...props} viewBox="0 0 24 24">
+    <path d="M19.8 4.2L16 8l-2-2-6 6-2-2-4 4v6h6l4-4-2-2 6-6-2-2 3.8-3.8c.6-.6 1.5-.6 2.1 0 .6.6.6 1.5 0 2.1zM5 18v-2.8l2-2 2 2-2 2H5z" />
+  </SvgIcon>
+);
+
+export const DaneelIcon = (props: SvgIconProps) => (
+  <SvgIcon {...props} viewBox="0 0 24 24">
+    <path d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm0 2c4.41 0 8 3.59 8 8s-3.59 8-8 8-8-3.59-8-8 3.59-8 8-8zm-1 6v4h2v-4h-2zm0 6v2h2v-2h-2z" />
+  </SvgIcon>
+);
+
+export const GiskardIcon = (props: SvgIconProps) => (
+  <SvgIcon {...props} viewBox="0 0 24 24">
+    <path d="M17 16l-4-4V8.82C14.16 8.4 15 7.3 15 6c0-1.66-1.34-3-3-3S9 4.34 9 6c0 1.3.84 2.4 2 2.82V12l-4 4H3v5h5v-3.05l4-4.2 4 4.2V21h5v-5h-4z" />
+  </SvgIcon>
+);
+
+export const CalvinIcon = (props: SvgIconProps) => (
+  <SvgIcon {...props} viewBox="0 0 24 24">
+    <path d="M20 8h-2.81c-.45-.78-1.07-1.45-1.82-1.96L17 4.41 15.59 3l-2.17 2.17C12.96 5.06 12.49 5 12 5s-.96.06-1.41.17L8.41 3 7 4.41l1.62 1.63C7.88 6.55 7.26 7.22 6.81 8H4v2h2.09c-.05.33-.09.66-.09 1v1H4v2h2v1c0 .34.04.67.09 1H4v2h2.81c1.04 1.79 2.97 3 5.19 3s4.15-1.21 5.19-3H20v-2h-2.09c.05-.33.09-.66.09-1v-1h2v-2h-2v-1c0-.34-.04-.67-.09-1H20V8zm-4 4v3c0 2.21-1.79 4-4 4s-4-1.79-4-4v-3c0-2.21 1.79-4 4-4s4 1.79 4 4z" />
+  </SvgIcon>
+);